+++ lib/src/provider/provider.dart
@@ -1,5 +1,6 @@
 import 'package:meta/meta.dart';
 import 'package:riverpie/src/notifier.dart';
+import 'package:riverpie/src/notifier/view_provider_notifier.dart';
 import 'package:riverpie/src/observer/observer.dart';
 import 'package:riverpie/src/provider/state.dart';
 import 'package:riverpie/src/ref.dart';
@@ -51,10 +52,25 @@ class Provider<T> extends BaseProvider<T> {
   }
 }
 
+/// A provider that holds a [BaseNotifier].
+/// Only notifiers are able to add listeners.
+abstract class BaseNotifierProvider<N extends BaseNotifier<T>, T>
+    extends BaseProvider<T> {
+  BaseNotifierProvider({super.debugLabel});
+
+  @internal
+  @override
+  NotifierProviderState<N, T> createState(
+    Ref ref,
+    RiverpieObserver? observer,
+  );
+}
+
 /// Use a [NotifierProvider] to implement a stateful provider.
 /// Changes to the state are propagated to all consumers that
 /// called [watch] on the provider.
-class NotifierProvider<N extends BaseNotifier<T>, T> extends BaseProvider<T> {
+class NotifierProvider<N extends BaseNotifier<T>, T>
+    extends BaseNotifierProvider<N, T> {
   @internal
   final N Function(Ref ref) create;
 
@@ -66,15 +82,47 @@ class NotifierProvider<N extends BaseNotifier<T>, T> extends BaseProvider<T> {
     Ref ref,
     RiverpieObserver? observer,
   ) {
-    final state = NotifierProviderState<N, T>(create(ref));
+    final notifier = create(ref);
+    final state = NotifierProviderState<N, T>(notifier);
 
     // ignore: invalid_use_of_protected_member
-    state.getNotifier().preInit(ref, observer);
+    notifier.preInit(ref, observer);
 
     return state;
   }
 
   ProviderOverride<T> overrideWithNotifier(N Function() notifier) {
     return ProviderOverride(this, NotifierProviderState(notifier()));
+  }
+}
+
+/// The [ViewProvider] is the only provider that can watch other providers.
+/// Its builder is similar to a normal [Provider].
+/// A common use case is to define a view model that depends on many providers.
+/// Don't worry about the [ref], you can use it freely inside any function.
+class ViewProvider<T> extends BaseNotifierProvider<ViewProviderNotifier<T>, T> {
+  @internal
+  final T Function(WatchableRef ref) create;
+
+  ViewProvider(this.create, {super.debugLabel});
+
+  @override
+  NotifierProviderState<ViewProviderNotifier<T>, T> createState(
+    Ref ref,
+    RiverpieObserver? observer,
+  ) {
+    final notifier = ViewProviderNotifier<T>(create);
+    final state = NotifierProviderState<ViewProviderNotifier<T>, T>(notifier);
+
+    notifier.preInit(ref, observer);
+
+    return state;
+  }
+
+  ProviderOverride<T> overrideWithBuilder(T Function(WatchableRef) builder) {
+    return ProviderOverride(
+      this,
+      NotifierProviderState(ViewProviderNotifier(builder)),
+    );
   }
 }