+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.7.0
 
+- feat: add `ViewProvider`, the only provider that can watch other providers
 - feat: add `exclude` parameter for `RiverpieDebugObserver`
 
 ## 0.6.0